+++ Editor/Scripts/AnimeToolboxConstants.cs
@@ -3,7 +3,7 @@
 namespace Unity.AnimeToolbox.Editor {
 
 internal static class AnimeToolboxConstants  {
-    private const string PACKAGE_NAME = "com.unity.anime-toolbox";
+    internal const string PACKAGE_NAME = "com.unity.anime-toolbox";
     public static readonly string PACKAGE_PATH = Path.Combine("Packages", PACKAGE_NAME);