+++ Editor/Scripts/Package/PackageLifecycle.cs
@@ -1,0 +1,15 @@
+﻿namespace Unity.AnimeToolbox.Editor {
+
+/// <summary>
+/// The lifecycle status of a package
+/// </summary>
+public enum PackageLifecycle {
+    INVALID,
+    EXPERIMENTAL,
+    PREVIEW,
+    PRERELEASE,
+    RELEASED,
+}
+
+}
+