+++ Editor/Scripts/PackageRequest/RequestInfo/PackageListRequestInfo.cs
@@ -3,13 +3,13 @@ using UnityEditor.PackageManager.Requests;  //ListRequest, AddRequest, etc
 using UnityEditor.PackageManager;           //PackageCollection
 
 namespace Unity.AnimeToolbox.Editor {
-internal class ListRequestInfo {
+internal class PackageListRequestInfo {
     internal readonly bool OfflineMode;
     internal readonly bool IncludeIndirectIndependencies;
     internal readonly Action<Request<PackageCollection>> OnSuccessAction;
     internal readonly Action<Request<PackageCollection>> OnFailAction;
 
-    internal ListRequestInfo(bool offlineMode, bool includeIndirectDependencies,
+    internal PackageListRequestInfo(bool offlineMode, bool includeIndirectDependencies,
         Action<Request<PackageCollection>> onSuccess, Action<Request<PackageCollection>> onFail)
     {
         OfflineMode = offlineMode;