+++ Editor/Scripts/PackageRequest/RequestInfo/PackageRemoveRequestInfo.cs
@@ -1,12 +1,12 @@
 ﻿using System;                               //Action
 
 namespace Unity.AnimeToolbox.Editor {
-internal class RemoveRequestInfo{
+internal class PackageRemoveRequestInfo{
     internal readonly string PackageName;
     internal readonly Action OnSuccessAction;
     internal readonly Action OnFailAction;
 
-    internal RemoveRequestInfo(string packageName,
+    internal PackageRemoveRequestInfo(string packageName,
         Action onSuccess, Action onFail)
     {
         PackageName = packageName;